+++ tests/test_codec.py
@@ -394,6 +394,21 @@ class TestPENMANCodec(object):
                          top='i2')
         assert encode(g, indent=None) == '(i2 / i :ARG0 (i / i))'
 
+    def test_encode_issue_67(self):
+        # https://github.com/goodmami/penman/issues/61
+        triples = [('h', ':instance', 'have-org-role-91'),
+                   ('a', ':instance', 'activist'),
+                   ('h', ':ARG0', 'a'),
+                   ('h', ':ARG2', 'a')]
+        assert encode(penman.Graph(triples, top='a')) == (
+            '(a / activist\n'
+            '   :ARG0-of (h / have-org-role-91)\n'
+            '   :ARG2-of h)')
+        assert encode(penman.Graph(triples, top='h')) == (
+            '(h / have-org-role-91\n'
+            '   :ARG0 (a / activist)\n'
+            '   :ARG2 a)')
+
     def test_parse_triples(self):
         assert codec.parse_triples('role(a,b)') == [
             ('a', 'role', 'b')]